Refactor out ColumnTypeDetector from CsvReader
