+++ core/src/main/java/tech/tablesaw/io/csv/CsvReader.java
@@ -25,7 +25,7 @@ import tech.tablesaw.api.ColumnType;
 import tech.tablesaw.api.Table;
 import tech.tablesaw.columns.AbstractParser;
 import tech.tablesaw.columns.Column;
-import tech.tablesaw.columns.strings.StringColumnType;
+import tech.tablesaw.io.ColumnTypeDetector;
 
 import javax.annotation.concurrent.Immutable;
 import java.io.BufferedReader;
@@ -39,62 +39,17 @@ import java.io.InputStreamReader;
 import java.io.Reader;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.concurrent.CopyOnWriteArrayList;
 
 import static tech.tablesaw.api.ColumnType.*;
 
 @Immutable
 public class CsvReader {
 
-    /**
-     * Consider using TextColumn instead of StringColumn for string data after this many rows
-     */
-    private static final int STRING_COLUMN_ROW_COUNT_CUTOFF = 50_000;
-
-    /**
-     * Use a TextColumn if at least this proportion of values are found to be unique in the type detection sample
-     *
-     * Note: This number is based on an assumption that as more records are considered, a smaller proportion of these
-     * new records will be found to be unique
-     *
-     * Sample calculation;
-     * 10 character string = 2 bytes * 10 + 38 extra bytes = 58; rounded up to 64 so it's a multiple of 8
-     *
-     * With dictionary encoding, we have 2*64 + 2*4 = 136 byte per unique value plus 4 bytes for each value
-     * For text columns we have 64 bytes per string
-     *
-     * So, if every value is unique, using dictionary encoding wastes about 70 bytes per value.
-     * If there are only two unique values, dictionary encoding saves about 62 bytes per value.
-     *
-     * Of course, it all depends on the lengths of the strings.
-     */
-    private static final double STRING_COLUMN_CUTOFF = 0.50;
-
-    /**
-     * Types to choose from. When more than one would work, we pick the first of the options. The order these appear in
-     * is critical. The broadest must go last, which is why String is at the end of the list. Any String read from
-     * a CSV will match string. If it were first on the list, you would get nothing but strings in your table.
-     *
-     * As another example, an integer type, should go before double. Otherwise double would match integers so
-     * the integer test would never be evaluated and all the ints would be read as doubles.
-     */
-    private List<ColumnType> typeArray =
-            Lists.newArrayList(
-                    LOCAL_DATE_TIME,
-                    LOCAL_TIME,
-                    LOCAL_DATE,
-                    BOOLEAN,
-                    SHORT,
-                    INTEGER,
-                    LONG,
-                    FLOAT,
-                    DOUBLE,
-                    STRING,
-                    TEXT);
+    private List<ColumnType> typeArrayOverrides = null;
 
     /**
      * Constructs a CsvReader
@@ -107,7 +62,7 @@ public class CsvReader {
      * These are the only types that the CsvReader can detect and parse
      */
     public CsvReader(List<ColumnType> typeDetectionList) {
-        this.typeArray = typeDetectionList;
+        this.typeArrayOverrides = typeDetectionList;
     }
 
     public Table read(CsvReadOptions options) throws IOException {
@@ -454,17 +409,7 @@ public class CsvReader {
     protected ColumnType[] detectColumnTypes(InputStream stream, CsvReadOptions options) {
 
         boolean header = options.header();
-        boolean useSampling = options.sample();
-
         int linesToSkip = header ? 1 : 0;
-
-        // to hold the results
-        List<ColumnType> columnTypes = new ArrayList<>();
-
-        // to hold the data read from the file
-        List<List<String>> columnData = new ArrayList<>();
-
-        int rowCount = 0; // make sure we don't go over maxRows
 
         CsvParser csvParser = csvParser(options);
 
@@ -475,117 +420,29 @@ public class CsvReader {
                 csvParser.parseNext();
             }
 
-            String[] nextLine;
-            int nextRow = 0;
-            while ((nextLine = csvParser.parseNext()) != null) {
-                // initialize the arrays to hold the strings. we don't know how many we need until we read the first row
-                if (rowCount == 0) {
-                    for (int i = 0; i < nextLine.length; i++) {
-                        columnData.add(new ArrayList<>());
-                    }
-                }
-                int columnNumber = 0;
-                if (rowCount == nextRow) {
-                    for (String field : nextLine) {
-                        columnData.get(columnNumber).add(field);
-                        columnNumber++;
-                    }
-                    if (useSampling) {
-                        nextRow = nextRow(nextRow);
-                    } else {
-                        nextRow = nextRowWithoutSampling(nextRow);
-                    }
-                }
-                rowCount++;
-            }
+            ColumnTypeDetector detector = typeArrayOverrides == null
+        	    ? new ColumnTypeDetector() : new ColumnTypeDetector(typeArrayOverrides);
+            return detector.detectColumnTypes(new Iterator<String[]>() {
+
+        	String[] nextRow = csvParser.parseNext();
+
+		@Override
+		public boolean hasNext() {
+		    return nextRow != null;
+		}
+
+		@Override
+		public String[] next() {
+		    String[] tmp = nextRow;
+		    nextRow = csvParser.parseNext();
+		    return tmp;
+		}
+        	
+            }, options);
         } finally {
             csvParser.stopParsing();
             // we don't close the reader since we didn't create it
         }
-
-        // now detect
-        for (List<String> valuesList : columnData) {
-            ColumnType detectedType = detectType(valuesList, options);
-            if (detectedType.equals(StringColumnType.STRING) && rowCount > STRING_COLUMN_ROW_COUNT_CUTOFF) {
-                HashSet<String> unique = new HashSet<>(valuesList);
-                double uniquePct = unique.size() / (valuesList.size() * 1.0);
-                if (uniquePct > STRING_COLUMN_CUTOFF) {
-                    detectedType = TEXT;
-                }
-            }
-            columnTypes.add(detectedType);
-        }
-        return columnTypes.toArray(new ColumnType[0]);
-    }
-
-    private int nextRowWithoutSampling(int nextRow) {
-        return nextRow + 1;
-    }
-
-    private int nextRow(int nextRow) {
-        if (nextRow < 10_000) {
-            return nextRow + 1;
-        }
-        if (nextRow < 100_000) {
-            return nextRow + 1000;
-        }
-        if (nextRow < 1_000_000) {
-            return nextRow + 10_000;
-        }
-        if (nextRow < 10_000_000) {
-            return nextRow + 100_000;
-        }
-        if (nextRow < 100_000_000) {
-            return nextRow + 1_000_000;
-        }
-        return nextRow + 10_000_000;
-    }
-
-    /**
-     * Returns a predicted ColumnType derived by analyzing the given list of undifferentiated strings read from a
-     * column in the file and applying the given Locale and options
-     */
-    private ColumnType detectType(List<String> valuesList, CsvReadOptions options) {
-
-        CopyOnWriteArrayList<AbstractParser<?>> parsers = new CopyOnWriteArrayList<>(getParserList(typeArray, options));
-
-        CopyOnWriteArrayList<ColumnType> typeCandidates = new CopyOnWriteArrayList<>(typeArray);
-
-        for (String s : valuesList) {
-            for (AbstractParser<?> parser : parsers) {
-                if (!parser.canParse(s)) {
-                    typeCandidates.remove(parser.columnType());
-                    parsers.remove(parser);
-                }
-            }
-        }
-        return selectType(typeCandidates);
-    }
-
-    /**
-     * Returns the selected candidate for a column of data, by picking the first value in the given list
-     *
-     * @param typeCandidates a possibly empty list of candidates. This list should be sorted in order of preference
-     */
-    private ColumnType selectType(List<ColumnType> typeCandidates) {
-        return typeCandidates.get(0);
-    }
-
-    /**
-     * Returns the list of parsers to use for type detection
-     *
-     * @param typeArray Array of column types. The order specifies the order the types are applied
-     * @param options CsvReadOptions to use to modify the default parsers for each type
-     * @return  A list of parsers in the order they should be used for type detection
-     */
-    private List<AbstractParser<?>> getParserList(List<ColumnType> typeArray, CsvReadOptions options) {
-        // Types to choose from. When more than one would work, we pick the first of the options
-
-        List<AbstractParser<?>> parsers = new ArrayList<>();
-        for (ColumnType type : typeArray) {
-            parsers.add(type.customParser(options));
-        }
-        return parsers;
     }
 
     private CsvParser csvParser(CsvReadOptions options) {
@@ -608,10 +465,4 @@ public class CsvReader {
         return format;
     }
 
-    /**
-     * Returns the list of types that specifies the order in which types are tested in the detection algorithm.
-     */
-    public List<ColumnType> getTypeArray() {
-        return typeArray;
-    }
 }
